requirements.txt
numpy==1.26.4
opencv-python==4.11.0.86
open3d==0.19.0
pandas==2.3.3
torch==2.8.0
torchvision==0.23.0
hydra-core==1.3.2
iopath==0.1.10
pillow==10.1.0
